Update to 1.15.2

+++ src/main/java/nl/parrotlync/discovoutlines/command/OutlinesCommandExecutor.java
@@ -1,6 +1,5 @@
 package nl.parrotlync.discovoutlines.command;
 
-import com.gmail.filoghost.holographicdisplays.api.HologramsAPI;
 import nl.parrotlync.discovoutlines.DiscovOutlines;
 import nl.parrotlync.discovoutlines.model.ReferencePoint;
 import nl.parrotlync.discovoutlines.util.ChatUtil;
@@ -9,8 +8,6 @@ import org.bukkit.Material;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
 import org.bukkit.command.TabExecutor;
-import org.bukkit.entity.ArmorStand;
-import org.bukkit.entity.EntityType;
 import org.bukkit.entity.Player;
 import org.bukkit.util.StringUtil;
 
@@ -48,7 +45,7 @@ public class OutlinesCommandExecutor implements TabExecutor {
 
                 if (args[1].equalsIgnoreCase("list")) {
                     List<ReferencePoint> points = DiscovOutlines.getInstance().getReferenceManager().getPoints();
-                    int pages = (int) Math.ceil((points.size() / 5) + 1);
+                    double pages = Math.ceil((double) (points.size() / 5) + 1);
                     int page = 1;
 
                     if (args.length == 3) {
